+++ Engine/Graphics/OpenGL/OGLGraphics.cpp
@@ -1154,7 +1154,7 @@ void Graphics::SetShaderParameter(StringHash param, const Matrix3x4& matrix)
 
 bool Graphics::NeedParameterUpdate(ShaderParameterGroup group, const void* source)
 {
-    if ((unsigned)shaderParameterSources_[group] == M_MAX_UNSIGNED || shaderParameterSources_[group] != source)
+    if ((unsigned)(size_t)shaderParameterSources_[group] == M_MAX_UNSIGNED || shaderParameterSources_[group] != source)
     {
         shaderParameterSources_[group] = source;
         return true;
@@ -1923,7 +1923,7 @@ void Graphics::FreeScratchBuffer(void* buffer)
         }
     }
     
-    LOGWARNING("Reserved scratch buffer " + ToStringHex((unsigned)buffer) + " not found");
+    LOGWARNING("Reserved scratch buffer " + ToStringHex((unsigned)(size_t)buffer) + " not found");
 }
 
 void Graphics::CleanupScratchBuffers()
